REST API uses ZOOM transactions
This changes the cluster REST API to use ZOOM transactions, where
the read/write operations during the same API request are processed
as an atomic transaction.

This is initial step that modifies the MidonetResource class to
provide access to a request-based transaction via the operation
filters. It also modifies most resources towards using transactions.
However, the following resources are not fully transactional and
may require future work:

 - BridgeResource
 - L2InsertionResource
 - IPAddrGroupResource
 - PoolMemberResource
 - VipResource
 - VtepResource
 - VtepBindingResource

Change-Id: I93c2247492e1b1cd2de4380130be6cbac8575e79
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>