Update patch set 11
Patch Set 11: Code-Review-1

(3 comments)

Some of the locks added to the API are redundant with each other. 

We just need locks on resources that are doing a get/update on any of the resources that have references to a Router Port or to a Router (which are the ones modified in the HM).  The Vtep for example touches port, but never a Router port.

Patch-set: 11
Label: Code-Review=-1