Update patch set 7
Patch Set 7:

I took a look at the latest failure:

mdts.tests.functional_tests.test_load_balancer.test_multi_member_loadbalancing

It is failing because the actual balancing, based on weight, doesn't look right:

Weights: [('10.0.1.2', 50), ('10.0.1.1', 25), ('10.0.1.3', 5)]
Hits: [('10.0.1.3', 20), ('10.0.1.2', 10), ('10.0.1.1', 5)]

(Thanks Xavi for putting in enough logs to see that)

I don't think this is related to this patch. I see other, unrelated patches failing on this test intermittently, and the failure doesn't *seem* to be related to the changes in this patch.

I'm ok with merging as is.

Patch-set: 7
Label: Code-Review=0