Fix race between chain updates
In l2insertion, when an insertion is added, it gets translated into
multiple rules in a chain. If multiple insertions are added to a port at
the same time, the generated rules may be deleted before the agent has
been able to load them. If this isn't handled gracefully, the agent will
mark the port as errored, which is undesirable.

This patch fixes this by simply ignoring the not found error. The
RuleState will never emit, so the chain will not complete until we
receive the updated chain, from which the missing rule will have been
removed.

Change-Id: Ia740990cf63733a541683f7d9e54d2fb249591d6
Signed-off-by: Ivan Kelly <ivan@midokura.com>