Cluster: introduce DeviceWatcher utility
A class that can be used to watch for changes on devices of a given
type, and perform a given action when one is changed. It allows
filtering events based on some provided conditions. This is just some
boilerplate on top of the ZOOM observers that cannot be done straight
using operators. 

Ref.: MNA-1049
Signed-off-by: Galo Navarro <galo@midokura.com>
Signed-off-by: Xavi León <xavi.leon@midokura.com>
Change-Id: Ibbe9b7bff9bf9567a7d3951c1bb084e460f77c05