Cluster: introduce DeviceWatcher utility
A class that can be used to watch for changes on devices of a given
type, and perform a given action when one is changed.  This is just some
boilerplate on top of the ZOOM observers that cannot be done straight
using operators.

Ref.: MNA-1049
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Ia66740a3db4b7ca08fc278c6eed8a67555843953

Cluster: allow filtering on DeviceWatchers

Although a client may be interested in watching a specific device, it
may not be interested in all changes. This patchs introduces a new
handler function to filter messages based on some condictions.

Change-Id: Ibbe9b7bff9bf9567a7d3951c1bb084e460f77c05
Signed-off-by: Xavi León <xavi.leon@midokura.com>