Handle race conditions the router mapper
This modifies the router mapper to handle situations where a
router child device (port, route, load balancer) was deleted before
being loaded from storage. The handling code ensures the following:

(a) If a port is deleted before being loaded (the error is
    NotFoundException, and the port state is not ready), the error
    is ignored and the port observable completes immediately.
    Otherwise, the error is allowed to complete with error the
    router observable.

(b) If a route emits any error (either before or after being
    loaded from storage), the error is ignored, logged and the
    route removed from the routing table.

(c) The behavior for load balancer is the same as for port.

Note: The subresources handled by the ObjectReferenceTracker
(mirror and chains) will be addressed in a subsequent patch.

Fixes: MNA-1062

Change-Id: If12515fbb95c1704aff22d6c5f0a41fd170865b0
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>