MDTS: Await pool members are active before L4LB test
This change awaits the pool members are active before sending
traffic via the load balancer. This because the letancy of a pool
member being marked as active when pool health monitoring is
enabled. Otherwise, it is possible that one or more pool members
are not active when traffic is sent via the load balancer, breaking
the test assumption of the pool members being hit.

Fixes: MNA-1090

Change-Id: I59c8f9f5323be65b4d1ed54928cd75f2bf6ad0eb
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>