Replicated tables: Move the ZK directory to NSDB
This is preparation for unifying the paths between the replicated
maps and the v5.0 models, which ensures that new replicated maps
use the same path model as the v5.0 topology/state data.

The objective of this migration is to have each replicated map
defined when building the backend for the classes that require
them (e.g. bridge, router), and the backend will ensure the
ZK paths are created for new objects. It will also offer the
capability of accessing them using the MidonetBackend rather than
the legacy DataClient.

Eventually this will allow us to delete the DataClient and break
the dependency between the midolman and midonet-cluster modules.

Change-Id: Ic8b84c2a707207d0344f719652229de3a1b7972e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>