Replicated tables: Introduce ZK state table storage
With this change, new replicated tables can be created from the
MidonetBackend. The current implementation allows state tables using
ZooKeeper as backend, usually leveraging the existing implementation
of the ReplicatedMap. Toward this, you must register a new state
table as:

  store.registerTable(clazz, key, value, name, provider)

Where:

  clazz:    Is the object class for which the table is registered
            in storage (e.g. classOf[Router])
  key:      Is the table key class (e.g. classOf[IPv4Addr])
  value:    Is the table value class (e.g. classOf[MAC])
  provider: Is the class that provides the implementation for the
            state table.
  name:     Is the name of the state table. The name must be
            unique for a given object class.

The state table provider class can be any class with a public
constructor that allows the injection using Guice of the ZK
directory where the state table is allowed to store the data.
The Guice injection also provides a ZK connection watcher if
needed by the state table implementation.

To retrieve an instance of a state table use:

  store.getTable[K,V](clazz, id, name): StateTable[K,V]

Where:

  K:     Is the table key class (e.g. IPv4Addr).
  V:     Is the table value class (e.g. MAC).
  clazz: Is the object class (e.g. classOf[Router])
  id:    Is the object identifier for which the table is created.
  name:  Is the state table name.

Change-Id: I04f56d05400f3ec2f0ffa871d9082a622136f4b3
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>