Drop RequestObserver, use better Rx idioms.
This class is a plain NoOp Observer.  Apart of having the wrong names,
it's using a very weird Rx idiom.  Replace with a .finallyDo
alternative.

Incidentally, this improves coverage reports as RequestObserver had
no unit tests and the classes that touch it were on a separate module,
which doesn't include them in the report.

Change-Id: Ie1574f0fabae5fba6675e6abe4afea78c298f15d
Signed-off-by: Galo Navarro <galo@midokura.com>