Ensure flow state is available after a redirect
When redirecting for service chain, the flow state will normally
only be sent to the host which has the first service port in the
chain. However, as services in the chain are traversed in the same order
on ingress and egress, this means that flow state may not be available
on egress, as the last service port may not be on any host which has
flow state.

FailOpen complicates this further, as if a service port has failed the
egress simulation may not even occur on the host with the last service
port, but rather the host with the last non-failed service port.

This patch resolve this by sending flow state to all service ports when
a packet is redirected.

Change-Id: Idbad9ff948b8bf994a979aa06429fb7304498147
Signed-off-by: Ivan Kelly <ivan@midokura.com>