Moving Pool mapping status to state store
This patch moves the health monitor-pool mapping to the state store.
The mapping status was previously stored together with the pool in
Zoom.

The equivalent of the previously-used "INACTIVE" mapping state is
now the absence of the corresponding state key.
We keep the INACTIVE value as other classes rely on it however.

API tests are also modified. This is because previously, the
default mapping status was ACTIVE. With the mapping in the state
store however, the absence of a state key is interpreted as
a PENDING_CREATE. As a consequence, we change the mapping
status in several tests to allow the modification of the pool.

Ref: MI-3

Change-Id: I1430eac814520c216c3627e549e11e2e7055d4d5
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>