Replicated tables: Add connection mgmt to backend
This adds the reactor, legacy ZK connection, and the connection
watcher to the MidonetBackend. This will be later needed to allow
the backend to create ZooKeeper-backed replicated maps, but will
also be used by other modules that require connection management,
such as the BGP routing handler.

This will eventually allow to remove the legacy ZooKeeper
connection module.

With this change the ZK connection and ZK connection watcher is
backed by the Curator connection.

Change-Id: Ieb81fcc32f59c45f330e6cee42bd5e486c7d9880
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>