MDTS/Sandbox: pull latest images available on CI
Latest step to automate the building of sandbox images. At the start of
the job, we pull the latest images available (from artifactory if
running on the CI). These images still won't be available outside of
the CI network, to be done in a later stage. Need to update the
midonet-sandbox revision as a latest change is needed for this
to properly work.

Change-Id: I1ef2a4851b8669a3c8f25e811c57bf15320bfff2
Signed-off-by: Xavi León <xavi.leon@midokura.com>