VPNaaS: Introduce container delegates and handlers
A container delegate specifies the cluster custom operations to be
executed when a new container is created, set up, teared down or
deleted. This accommodates containers of different types having to
perform certain operations when a container state changes (e.g.
modify objects in the topology, etc.).

A container handler implements at the agent the operations to
setup and teardown a container namespace, local veth interface,
the configuration and execution of the programs associate with the
container.

This patch introduces a new annotation and interfaces for service
delegates and handlers. At startup the container managament at both
the cluster and agent will load all service delegate and handler
implementation, respectively, from the current class path and apply
their actions when the containers configuration and status has
changed.

Change-Id: I88985dc048e7c2b5b606207a24e14a966bfb594c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>