VPNaaS: A container handler provider for the agent
The container provider scans the current classpath for
implementations of the ContainerHandler trait annotated with the
Container annotation. This allows the agent to load any container
handler provided in the current classpath in the package:

  org.midonet.midolman.topology.containers

Multiple implementations of a container handler are distinguished
by the name and version specified in the Container annotation. The
handler name must match the container type specified in the
service container group.

Change-Id: I6af69c80e948f906b8c4d620b0f1b94390db6b3e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>