VPNaaS: A container delegate provider for the cluster
This provides a container delegate provider for the cluster,
similar to the container handler provider for the agent. Toward
this end, common code has been moved to a single ContainerProvider
abstract class that receives the package prefix and class type as
arguments. The handler and delegate provides inherit this class,
and specify the prefix and the injector to use for creating
instances.

Change-Id: Ic328f153f8cd500f03d88da091e51b8f1eb5e4e3
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>