VPNaaS: Simplify the virtual-to-physical mapper
This change intends to accommodate the addition of the container
service to the agent. The service will be added to the virtual-to-
physical mapper, in addition to the existing VXLAN port mapping
service. The rationale for this is that the VTPM can intercept
port bindings corresponding to service containers, directing the
container service to start or stop the appropriate container.

To this end, the current structure of the VTPM has been simplified
considerably, removing all legacy data structures, and bringing the
code in line with the V5 architecture. The VTPM is now an abstract
service rather than an actor (this is needed only to launch the
child services such as the VXLAN port mapping service, container
service etc.), and all updates are exposed as observables taken
directly from the virtual topology and only protected with
error-handling code. The tryAsk method has been removed since it
wasn't used.

The consumers of VTPM notifications (mainly the HostRequestProxy
and the DatapathController) have been updated as reactive actors
and now use subscriptions to the VTPM observables to receive
host and tunnel-zone updates.

The VTPM redirector is removed, since it was only needed to bridge
the legacy code with the new architecture.

Change-Id: I4397ca277e67b2ab466feeb86827efa85004dbac
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>