Build: fix circular dependency
The agent sim. tests were depending on the cluster in order to reuse
L2Insertion translations.  These have no dependencies on midonet-cluster
or midolman, so they get pushed down to NSDB.  Eventually it's likely
that we'll move all here, in order to be reused by simulation tests to
build topologies.

Change-Id: Ib61980fe86948459d7141c3bfd2fe9cb483fb2a5
Signed-off-by: Galo Navarro <galo@midokura.com>