VPNaas: Introduce a container mapper for the agent
This introduces a mapper that monitors the notification for a
specified host for changes in port bindings that are attched to
containers. The mapper implements the RX OnSubscribe interface,
and an observable create on the mapper will emit notifications
for the lifecycle of the container: created, changed and deleted.

The container notifications include the container host, port,
interface name, container identifier and the container configuration
identifier.

The mapper emits notifications for the following changes:
(a) port binding updates for the monitored host
(b) changes for a given binding, including whether the binding
    relationship to a container has changed
(c) changes to the configuration of a container

The mapper does not validate whether a service container has a
valid configuration, and it always emit a notification as long
as the new container configuration is distinct from the previous
one. The mapper does not emit notifications when other properties
of a container have changed (e.g. the container group, etc.)

The mapper filters any errors emitted on the observables for a
port binding and container.

The mapper ensures the cleanup of the containers by emitting
cleanup notifications to all subscribers for the following
situations:
(a) the `complete` method is called (e.g. when the agent shuts down)
(b) the host is deleted or emits an error
(c) a port binding is deleted or emits an error
(d) a container is deleted, removed from a port binding or emits
    an error

Change-Id: I3666d8ffed284ab2453ddb37bcd08044d3385ff0
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>