Accessing JMX from within MDTS
This patch adds a JMXMonitor class that can be used to access the metrics
from for example zookeeper, the cluster or midolmans over RMI. The connection
to the remote JVM is made using Java's own classes. For this to be possible the
JPype module needs to be installed.

Change-Id: Iea7f5fb3c2318cf786174fb553afcac2f40316ba
Signed-off-by: Huub Hendrikx <huub@midokura.com>
(cherry picked from commit a47b5291c288fae1a6818f5d36daa99ad4475869)