Agent: Router can peek Bridge's ARP cache
With this patch, the router simulation gains the ability to peer into
the ARP cache of the bridge where the output port's peer port is
connected, thus avoiding ARPing when the bridge knows the answer
(either because we're ARPing for an interior port connected to the
bridge or the bridge has been pre-seeded with the ARP table entry).
This further avoids duplication of data and churn on Zookeeper.

Change-Id: Ic4d4def163e0863776c61c31873c810ecfc9183d
Signed-off-by: Duarte Nunes <duarte@midokura.com>