VPNaaS: A container management service for the agent
This introduces a new container management service in the agent,
which is initialize with the VXLAN port mapping service in the
virtual-to-physical mapper. The service creates a single observable
on a container mapper for the current host and monitors the
topology data for container created for the current host.

When a new container is notified, the service uses the container
handler provider to create a new container handler instance.
The container handler classes are loaded from the classpath
package as specified in the provider, and the selected class is
done based on the service type from the service container group.

The service calls the created, updated, and deleted methods of the
container handler instance, according to the notifications sent by
the mapper observable.

For every container, the service monitors the health observable
of the container handler, and updates the container status in
storage accodingly. The service also ensures than the container
status is updated when the container handler returns an error or
throws an exception, or when the container observable terminates
unexpectedly.

Upon shutdown, the service calls the deleted method of all running
containers and clears their status in the backend. The service
waits a configurable time interval for the containers to stop
gracefully.

The service requires a single-thread executor on which to schedule
the container notificatons. In order to ensure an ordered sequence
of container events, all container changes are scheduled on this
executor. It should be expected that the executor may be blocked
by I/O operations, therefore do not use an executor that should
never be blocked for long time intervals, such as the VT executor.

The current implementation ensures a container notification queue
of up to 0x1000 notifications. This should be increased if the
container handlers may block for long periods of time, or it is
expected a high churn rate for the containers.

Change-Id: I6d3d8f273d50416899091e898249fdf4e0c9ae3e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>