Agent: Add encap/decap mechanism
This patch adds a vxlan encap and decap mechanism to the agent, so packets
can be (d)encapsulated during a simulation.

This is done by recirculating the packets through the host IP stack. A
veth pair endpoint is connected to the datapath and an additional
vxlan tunnel port is created. For encapsulation, the agent xmits a
packet through the vxlan tunnel directed at the host endpoint of the
veth pair, which causes the packet to re-enter the datapath through a
normal netdev port. For deencapsulation, the agent xmits a packet
through the datapath endpoint of the veth pair with the vxlan
destination port matching that of the recirculation vxlan tunnel port,
which causes the packet to re-enter the datapath through it.

Encapsulation is requested via the PacketContext. Given the outer
headers, the PacketContext will replace all the appropriate fields so
that it is the outer packet that the simulation will henceforth
consider.

Upon flow translation, the FlowTranslator prepares two flows. One that
matches on the packet, applies transformations and emits it out of the
tunnel port directed at the veth endpoint. A unique value is associated
with the tunnel header's TOS and TTL fields. A second flow matches on these
fields, resets fields such as the IP source and destination, and applies
transformations that were added during the simulation of the outer packet,
finally outputting the packet through the final destination port. These
flows are installed in reverse order by the FlowProcessor. The normal
PacketExecutor is made to ignore recirculating packets and a new one,
executed after the FlowProcessor, executes the packet, which will be
recirculated and will match the first flow we created.

Dencapsulation follows a similar procedure.

Right now only one encap/decap operation is permitted during a
simulation, but this restriction can be lifted by replacing scalar
fields in the PacketContext with lists.

Flow tracing at the egress is disabled for recirculated packets.

Change-Id: I6180f153c30bdaa5674dab29256b09cd4c0bd4e1
Signed-off-by: Duarte Nunes <duarte@midokura.com>