Loadbalancer mdts fix
This patch ensures that a different port is used for each backend
request in test_load_balancer.py. This allows requests to non sticky
vips to reach different backends.

This patch also lowers the number of requests made to shorten tests
while keeping the number high enough so that with very high
probability all tests pass.

Change-Id: I3146bedc43f2740ef439ddcfd913642c721ee544
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>