REST API includes IDs for learned routes
The learned routes do not have an identifier assigned in storage
for efficiency reasons (they are also not required for routing).
However, the API must provide an identifier in the route DTOs.
Currently, the API uses the same identifier for all learned routes
which causes clients, such as the midonet-cli to understanding it
as the same route.

With this change, learned routes returned by the API receive
identifiers that are computed deterministically from the route
data.

Change-Id: Iabfbbb76ce8db3efb434f26586e6d715dda83620
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>