Allow multiple tunnel zone members with the same IP
When reinstalling a node, it will generate a new UUID. When you add this
node to a tunnel zone, if you haven't removed the previous entry you
will end up with a duplicate IP.

As tunnelzone members are used in a UUID->IP lookup, duplicate IPs are
not a problem. However, previously we only allowed a single node to have
an IP, not purposefully, but because members were added to a Set, and
the object(HostConfig) in the set only looked at the IP for hashCode and
equals.

With this change, we look at the IP and the UUID in HostConfig#hashCode
and HostConfig#equals, so that we can have multiple members with the
same IP.

Change-Id: I0c890cc35c38ae16ba380d04bb3323b7cf8923fa
Signed-off-by: Ivan Kelly <ivan@midokura.com>