API: guard against inconsistent bindings
When a port with a binding is removed but the corresponding data in the
binding isn't cleaned, we returned ids of non existing hosts in the host
bindings dtos.  The clients might try to fetch these ports, and fail.

Similarly, a deletion of a binding might hit a port that is not in ZK.
If this happens, the port zk manager throws a not found exception.

This patch fixes these two problems: bindings are filtered so that only
those relative to existing ports are returned.  Additionally, the
deletion of a binding will only try to update the port when it does
exist in ZK.

Ref.: MN-3937, MNA-1044
Change-Id: Ie80bd9308e3eca357a0695d5c1fe8fbdebb08c2a
Signed-off-by: Galo Navarro <galo@midokura.com>