Update patch set 5
Patch Set 5:

There are examples on how to use the cli at docs.midokura.com. For instance for VTEPs:
http://docs.midokura.com/docs/latest/operation-guide/content/cli_remove_vtep_binding.html

Shouldn't we update these docs to also say how to manipulate containers? Or perhaps we should just create a ticket for that and wait until this feature has stabilized.

Patch-set: 5
Label: Code-Review=0