L4LB: Port routes should not include the router ID
In the ZOOM topology model, PORT routes only have referential
bindings to ports, not to the router. Otherwise, the same route
will be appear twice when listing the router routes.

This updates the L4LB haproxy health monitor to ensure the port
route added for the haproxy namespace is not added to the router.

It also updates the import order, and removes some Scala warnings.

Change-Id: I6850317749acf6d529b7c739cad184ff8c4e0258
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>