Fixing ObservablePathChildrenCache + sandbox CR bump
This patch ensures that subscribing to an OnSubscribeToPathChildren
and adding/removing children are performed atomically. This guarantees
that the map childStreams is accessed under mutual exclusion (the call
method may be called on a different thread than the ZooKeeper
notification thread).

More importantly, it ensures that subscribers receive all children
exactly once (subscribers do not miss notifications nor receive
children twice).

This patch also updates the sandbox CR so that it points to the
latest version.

Change-Id: Ic8d522762ae358b4fc9c209b3ad4aea77bfef76b
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>