VPNaaS: A port group tracker for container scheduler
This implements an ObjectTracker for a specified port group. The
class monitors the changes in the membership of the port group,
and exposes a stream of updates with the hosts and their container
service status where the active ports in the group are bound.

Internally, the class uses a ports CollectionTracker to receive
the updates for the active and bound ports in the group, and then
uses these updates to request to a hosts CollectionTracker the
updates for the corresponding hosts.

The tracker consolidates all updates in updates that include all
hosts elegible for container scheduling.

Change-Id: I3421c0c63a033bb14e8985ef4ec54c9fb06a1ef3
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>