VPNaaS: IPSec container is a container handler
The IPSec service container implements the container handler trait,
such that it is loaded and called by the agent container service.

The patch also enhances the robustness of the VPN helper script
and command calls, such that the container cleans-up properly upon
shutdown or failure. The container also atempts to clean the
file system and container namespace, if previously left in an
unclean state.

Change-Id: I6449f84dd59ba61af9b8a72950797fd45b07cbfb
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>