Conntrack: only tag flows when really necessary
The agent is over tagging based on conntrack keys. This
produces extra flow invalidations and is expensive.

Conntrack key querying can happen in different circumstances:

1.- A genuine, regular, forward flow. We query the conntrack table
and find nothing. We generate a conntrack key based on the egress
device (see trackConnection(), below) and install it on the table.

2.- A bogus forward flow. The flow is actually a return flow, but the
conntrack key has, for some reason, not arrived to this host yet.
We want to invalidate the resulting flow upon later arrival of
the key.

3.- A return flow. We query the conntrack table and find the conntrack
key for this flow.

Case #2 needs tag-based flow invalidation. Cases #1 and #3 do not need
it. There is no event related to the conntrack key (including
expiration, because a longer lifetime is actually desirable) that would
render the flow invalid.

Therefore, isForwardFlow must tag only if its result is true, based on
the possibility that the result is bogus and the conntrack key is due
to arrive at a later time.

Change-Id: Ib41d0648ad454bf6d44838fcb67b3deaa6ea9190
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>