VPNaaS: IPsec container delegate initial implementation
This initial implementation provides the means to create and remove a
binding when containers are allocated and deallocated to hosts. This
patch anotates the class with the Container annotation so the container
service can load it when a container of type "IPSEC" needs to do some
actions dictated by the scheduler. Failover actions (onUp and onDown)
will be added on a later patch.

Change-Id: I312a1134c3aa500905968483f5659aae4b56b628
Signed-off-by: Xavi León <xavi.leon@midokura.com>