VPNaaS: IPsec container delegate initial implementation
This initial implementation provides the means to create and remove a
binding when containers are allocated and deallocated to hosts. This
patch also introduces a factory class to instantiate delegates
depending on the type of container (which will be superseeded when
rebased on top of another patch that provides guice injection for
container delegates).

Change-Id: I312a1134c3aa500905968483f5659aae4b56b628
Signed-off-by: Xavi León <xavi.leon@midokura.com>