VxGW: don't recalculate FloodingProxy needlessly
This patch makes sure that the FloodingProxyManager only recalculates
the flooding proxy after a host changes if the diff contains relevant
changes.  For example, if we change the weight of a host or its
memberships to tunnel zones this requires recalculation.  But if we add
a binding to a host, we don't need to.

Change-Id: Id6cec9e728fbe44b843fcbba4656bc8420fe564f
Signed-off-by: Galo Navarro <galo@midokura.com>