Agent: Flow actions now inform the flow mask
Some flow actions, namely set key flow actions, may only be specified
when some fields of the flow match are exact-matched.

For example, to set a TCP flow key, the network protocol of the IP
flow key must be exact-matched and, by transitivity, so must be the
ether type.

To support older kernels (i.e., 3.13), setting an IP flow key also
requires an exact match on the IP protocol field.

Change-Id: I42956394aba0792534ee4b976b03d3e2fc84e7a7
Signed-off-by: Duarte Nunes <duarte@midokura.com>