[WIP] VPNaaS: Service container scheduler
This uses the scheduling algorithm from:
https://review.gerrithub.io/#/c/253057/
where the scheduling is done at the container level. The host
selector is still provided by a common factory to ensure that
we use a common selector when needed (e.g. a single anywhere host
selector).

This patch also adapts the scheduler to use observable
notifications rather than futures, toward the end of having the
subscriptions managed by the subscription to a single observable.

Switching to observables also ensures that we can add a feedback
observable loop that triggers a rescheduling when the status of
a container changes or the timeout for a scheduling operations
has expired. All notifications are handled on a single thread
for easy synchronization.

The scheduling for a container is a state machine, where the
possible states are Down, Scheduled and Up. Each state stores
per state information, such as the value of the service container
and service container group for which the scheduling was made,
the host where the container was scheduled, subscriptions to
the timeout timer observables, etc. The transition to a new state
follows an external notification, such as the available hosts have
changed, the container status has changed, etc, and the action
taken depends on the current scheduler state.

Note: The observers of this scheduler may not block the emitted
notifications for significant amount of time. The call to any I/O
operation should be rescheduled on a separate thread. The
observables internal to this scheduler do not use backpressure
buffers.

Change-Id: Ib781a1d85c6992fb5313b32c8e61467288a1f259
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
Signed-off-by: Xavi León <xavi.leon@midokura.com>
Signed-off-by: Galo Navarro <galo@midokura.com>