Only add necessary routes in VPN namespace
If we set a default route in the VPN namespace and the ike connection
hasn't been established (and SAs set up), then any outbound packet will
be pushed back to the router, and come back in to the namespace, then be
sent out to the router, ad infinitum. To avoid this, just set the routes
we need when a ipsec connection is added.

Change-Id: I282f6fbf9d27bbc7e121cea8004496beb3b3a4d6
Signed-off-by: Ivan Kelly <ivan@midokura.com>