Fix json property for ipsec_site_connection
Neutron expects a specific json property (ipsec_site_conn) which was 
changed and the plugin is not compatible anymore as it makes calls based
on the name of this json properties. As changing this implies doing 
changes on the plugin on master upstream and the backport to kilo, it 
seems easy to just change the json property in here and rename the var 
for internal purposes.

Change-Id: I27bc9665cf0370a91e518ff390d49e231b0eca67
Signed-off-by: Xavi León <xavi.leon@midokura.com>